+++ R/calculate_DIF_table.R
@@ -1,18 +1,17 @@
 #' Calculate DIF statistic per item for multiple groups
 #'
 #' This function calculates DIF statistics for multiple groups using the
-#' `difR` package. The function uses the `difGenLogistic()` function for
-#' generalized logistic regression and the `difGMH()` function for the
-#' generalized Mantel-Haenszel test. The function returns a table with
-#' DIF statistics per item.
+#' `difR` package. The function uses the `genDichoDif()` function for
+#' generalized logistic regression and the generalized Mantel-Haenszel test.
+#' The function returns a table with DIF statistics per item.
 #'
 #' @param data Data frame with dichotomous item responses and age
 #' @param items Character vector with item names
 #' @param group Factor or character vector with group names. Must have same
 #' length as the number of rows in the data frame.
-#' @param score Data frame with item parameters
-#' @param levels Optional. Character vector with group names. If omitted, the function
-#' will use the unique values of the group variable.
+#' @param score Data frame with item parameters. Optional.
+#' @param focal.names Optional. Character vector with group names.
+#' If omitted, the function will use the unique values of the group variable.
 #' @param thr Optional. Threshold for DIF classification using the MH-statistics.
 #' If omitted, the function will use the threshold from the `difGMH()` function.
 #' @details
@@ -21,74 +20,68 @@
 #' @return data frame with DIF statistics per item
 #'
 #' @examples
-#' \dontrun{
-#' # For some reason, the following example does not work with the
-#' # version 5.1 of difR. The example is kept here for future reference.
-#' # We get: Error in n.ppk * diag(rk) - rk %*% t(rk) : non-conformable arrays
-#'
-#' # DIF by sex, Dutch SMOCC data
 #' library(dscore)
 #' library(childdevdata)
 #' data <- childdevdata::gcdg_nld_smocc
-#' items <- intersect(get_itemnames(ins = "ddi", order = "indm"), colnames(data))
-#' score <- dscore(data = data, items = items, xname = "agedays", xunit = "days",
-#' metric = "logit")[["d"]]
-#' group <- as.character(data$sex)
+#' items <- intersect(dscore::get_itemnames(ins = "ddi", order = "indm"), colnames(data))
 #'
-#' # Remove rows with fewer than 3 observations
-#' # Remove missing values from group and score
-#' responses <- as.matrix(data[, items])
-#' sparse <- rowSums(!is.na(responses)) < 3
-#' idx <- !(is.na(score) | is.na(group) | sparse)
-#' responses <- responses[idx, ]
-#' group <- group[idx]
-#' score <- score[idx]
+#' # Method 1: Use a numeric group variable
+#' group <- ifelse(data$sex == "Male", 1, 0)
+#' focal.names <- 1
 #'
-#' library(difR )
-#' LR <- difLogistic(Data = responses,
-#'   group = group,
-#'   focal.names = unique(group),
-#'   match = score)
-#' MH <- difMH(Data = responses,
-#'   group = group,
-#'   focal.names = unique(group),
-#'   match = score)
+#' # Method 2: Use a character group variable
+#' group <- data$sex
+#' focal.names <- "Male"
+#'
+#' dt <- calculate_DIF_table(data = data, items = items, group = group,
+#'                          focal.names = focal.names)
+#'
 #' }
 #' @export
-calculate_DIF_table <- function(data, items, score, group,
-                                levels = NULL, thr = NULL) {
+calculate_DIF_table <- function(data, items, group, score = "score",
+                                focal.names = NULL, thr = NULL,
+                                p.adjust.method = "holm",
+                                method = c("GMH", "genLogistic")) {
 
   stopifnot(is.data.frame(data))
   stopifnot(length(items) > 0L)
-  stopifnot(length(score) == nrow(data))
+  stopifnot(length(score) == 1L || length(score) == nrow(data))
   stopifnot(length(group) == nrow(data))
 
-  group <- as.character(group)
-  if (is.null(levels)) {
-    levels <- unique(group)
+  # remove missing group values
+  idx <- !is.na(group)
+  data <- data[idx, ]
+  group <- as.character(group[idx])
+
+  if (is.null(focal.names)) {
+    focal.names <- unique(group)
   }
 
-  # -- Using difR::difGenLogistic() for multiple groups
-  LR <- difGenLogistic(Data = as.matrix(data[, items]),
-                       group = group,
-                       focal.names = levels,
-                       match = score,
-                       p.adjust.method = "holm")
+  # check length of focal.names
+  stopifnot(length(focal.names) == 1L ||
+              length(focal.names) == length(unique(group)))
 
-  # Derive Zumbo and Jodoin classifications from LR
-  LR_zumbo <- cut(LR$deltaR2,
-                  breaks = c(-Inf, 0.13, 0.26, Inf),
-                  labels = c(LETTERS[1:3]))
-  LR_jodoin <- cut(LR$deltaR2,
-                   breaks = c(-Inf, 0.035, 0.07, Inf),
-                   labels = c(LETTERS[1:3]))
+  # MH-test for uniform DIF
+  if ("GMH" %in% method) {
+    MH <- genDichoDif(Data = as.matrix(data[, items]),
+                      group = group,
+                      focal.names = focal.names,
+                      p.adjust.method = p.adjust.method,
+                      method = c("GMH"))
+  } else {
+    MH <- NULL
+  }
 
-  # -- Using difR::difGMH() for multiple groups
-  MH <- difGMH(Data = as.matrix(data[, items]),
-               group = group,
-               focal.names = levels,
-               match = score,
-               p.adjust.method = "holm")
+  if ("genLogistic" %in% method) {
+    # LR-test: unif + nunif
+    LR <- genDichoDif(Data = as.matrix(data[, items]),
+                      group = group,
+                      focal.names = focal.names,
+                      p.adjust.method = p.adjust.method,
+                      method = c("genLogistic"))
+  } else {
+    LR <- NULL
+  }
 
   # Derive DIF classification from MH using threshold thr
   if (is.null(thr)) {
@@ -98,6 +91,14 @@ calculate_DIF_table <- function(data, items, score, group,
   tmp <- data.frame(item = items,
                     MH_stat = MH$GMH,
                     MH_DIF = MH_DIF)
+
+  # Derive Zumbo and Jodoin classifications from LR
+  LR_zumbo <- cut(LR$deltaR2,
+                  breaks = c(-Inf, 0.13, 0.26, Inf),
+                  labels = c(LETTERS[1:3]))
+  LR_jodoin <- cut(LR$deltaR2,
+                   breaks = c(-Inf, 0.035, 0.07, Inf),
+                   labels = c(LETTERS[1:3]))
 
   # create item table sorted by DIF
   dif_table <- data.frame(num = 1:length(items),
@@ -112,4 +113,5 @@ calculate_DIF_table <- function(data, items, score, group,
                              "LR_stat", "LR_deltaR2", "LR_zumbo", "LR_jodoin")]
 
   return(dif_table)
+
 }